docker/event-generator/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
faker==25.0.0
azure-eventhub==5.11.4
